refactor: split schedules.rs (1,324 LOC) into focused submodules [OPE-262]
Split monolithic data/schedules.rs into a modular directory:
- mod.rs: types and public API coordination
- queries.rs: page loading, catalog building, and view construction
- mutations.rs: save, toggle, and delete operations
- tests.rs: comprehensive test suite (20 tests)

All public API preserved. All 24 schedule-related tests pass.
